# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/discovery/feed": {
      "get": {
        "operationId": "DiscoveryController_feed",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "default": "global",
              "type": "string",
              "enum": [
                "global",
                "following"
              ]
            }
          },
          {
            "name": "genre",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedPageDto"
                }
              }
            }
          }
        },
        "tags": [
          "discovery"
        ]
      }
    },
    "/v1/discovery/users/{userId}": {
      "get": {
        "operationId": "DiscoveryController_byUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "recent",
              "type": "string",
              "enum": [
                "recent",
                "score-desc",
                "score-asc"
              ]
            }
          },
          {
            "name": "genre",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "rock",
                "pop",
                "hip-hop",
                "r-b",
                "jazz",
                "electronic",
                "metal",
                "indie",
                "classical",
                "country",
                "reggae",
                "funk",
                "soul",
                "punk",
                "blues",
                "folk",
                "k-pop",
                "mpb",
                "samba",
                "sertanejo",
                "pagode"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedPageDto"
                }
              }
            }
          }
        },
        "tags": [
          "discovery"
        ]
      }
    },
    "/v1/discovery/users/{userId}/stats": {
      "get": {
        "operationId": "DiscoveryController_userStats",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatsDto"
                }
              }
            }
          }
        },
        "tags": [
          "discovery"
        ]
      }
    },
    "/v1/discovery/top-albums": {
      "get": {
        "operationId": "DiscoveryController_topAlbums",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "genre",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "rock",
                "pop",
                "hip-hop",
                "r-b",
                "jazz",
                "electronic",
                "metal",
                "indie",
                "classical",
                "country",
                "reggae",
                "funk",
                "soul",
                "punk",
                "blues",
                "folk",
                "k-pop",
                "mpb",
                "samba",
                "sertanejo",
                "pagode"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopAlbumsPageDto"
                }
              }
            }
          }
        },
        "tags": [
          "discovery"
        ]
      }
    },
    "/v1/discovery/albums/{albumId}/reviews": {
      "get": {
        "operationId": "DiscoveryController_albumReviews",
        "parameters": [
          {
            "name": "albumId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumReviewsPageDto"
                }
              }
            }
          }
        },
        "tags": [
          "discovery"
        ]
      }
    },
    "/v1/discovery/albums/{albumId}/stats": {
      "get": {
        "operationId": "DiscoveryController_albumStats",
        "parameters": [
          {
            "name": "albumId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumStatsDto"
                }
              }
            }
          }
        },
        "tags": [
          "discovery"
        ]
      }
    },
...
    "/v1/users/{userId}/follow": {
      "post": {
        "operationId": "FollowsController_follow",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowStatsDto"
                }
              }
            }
          }
        },
        "tags": [
          "follows"
        ]
      },
      "delete": {
        "operationId": "FollowsController_unfollow",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowStatsDto"
                }
              }
            }
          }
        },
        "tags": [
          "follows"
        ]
      }
    },
    "/v1/users/{userId}/follow-stats": {
      "get": {
        "operationId": "FollowsController_stats",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowStatsDto"
                }
              }
            }
          }
        },
        "tags": [
          "follows"
        ]
      }
    },
    "/v1/users/{userId}/followers": {
      "get": {
        "operationId": "FollowsController_followers",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUsersPageDto"
                }
              }
            }
          }
        },
        "tags": [
          "follows"
        ]
      }
    },
    "/v1/users/{userId}/following": {
      "get": {
        "operationId": "FollowsController_following",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUsersPageDto"
                }
              }
            }
          }
        },
        "tags": [
          "follows"
        ]
      }
    },
...
  },
  "info": {
    "title": "Track by Track API",
    "description": "Contrato da API — fonte de tipos para o frontend (ver seção 2.2 do CLAUDE.md)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FeedPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "userId": {
                  "type": "string"
                },
                "userDisplayName": {
                  "type": "string"
                },
                "userAvatarUrl": {
                  "type": "string"
                },
                "albumId": {
                  "type": "string"
                },
                "albumName": {
                  "type": "string"
                },
                "albumArtist": {
                  "type": "string"
                },
                "albumImageUrl": {
                  "type": "string"
                },
                "averageScore": {
                  "type": "number"
                },
                "ratedTracks": {
                  "type": "number"
                },
                "totalTracks": {
                  "type": "number"
                },
                "createdAt": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                },
                "badge": {
                  "type": "string",
                  "enum": [
                    "new",
                    "updated",
                    null
                  ],
                  "nullable": true
                }
              },
              "required": [
                "id",
                "userId",
                "userDisplayName",
                "albumId",
                "albumName",
                "albumArtist",
                "averageScore",
                "ratedTracks",
                "totalTracks",
                "createdAt",
                "updatedAt",
                "badge"
              ]
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "perPage": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "totalPages": {
                "type": "number"
              },
              "nextCursor": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "page",
              "perPage",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UserStatsDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "averageScore": {
            "type": "number"
          },
          "tracksRated": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "averageScore",
          "tracksRated"
        ]
      },
      "TopAlbumsPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "albumId": {
                  "type": "string"
                },
                "albumName": {
                  "type": "string"
                },
                "albumArtist": {
                  "type": "string"
                },
                "albumImageUrl": {
                  "type": "string"
                },
                "averageScore": {
                  "type": "number"
                },
                "ratingsCount": {
                  "type": "number"
                }
              },
              "required": [
                "albumId",
                "albumName",
                "albumArtist",
                "averageScore",
                "ratingsCount"
              ]
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "perPage": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "totalPages": {
                "type": "number"
              },
              "nextCursor": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "page",
              "perPage",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "AlbumReviewsPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "rankingId": {
                  "type": "string"
                },
                "userId": {
                  "type": "string"
                },
                "userDisplayName": {
                  "type": "string"
                },
                "userAvatarUrl": {
                  "type": "string"
                },
                "reviewText": {
                  "type": "string",
                  "nullable": true
                },
                "averageScore": {
                  "type": "number"
                },
                "createdAt": {
                  "type": "string"
                }
              },
              "required": [
                "rankingId",
                "userId",
                "userDisplayName",
                "reviewText",
                "averageScore",
                "createdAt"
              ]
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "perPage": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "totalPages": {
                "type": "number"
              },
              "nextCursor": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "page",
              "perPage",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "AlbumStatsDto": {
        "type": "object",
        "properties": {
          "albumId": {
            "type": "string"
          },
          "averageScore": {
            "type": "number"
          },
          "ratingsCount": {
            "type": "number"
          },
          "topFavoriteTracks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "trackId": {
                  "type": "string"
                },
                "trackName": {
                  "type": "string"
                },
                "percentage": {
                  "type": "number"
                }
              },
              "required": [
                "trackId",
                "trackName",
                "percentage"
              ]
            }
          },
          "topWorstTracks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "trackId": {
                  "type": "string"
                },
                "trackName": {
                  "type": "string"
                },
                "percentage": {
                  "type": "number"
                }
              },
              "required": [
                "trackId",
                "trackName",
                "percentage"
              ]
            }
          }
        },
        "required": [
          "albumId",
          "averageScore",
          "ratingsCount",
          "topFavoriteTracks",
          "topWorstTracks"
        ]
      },
...
      "FollowStatsDto": {
        "type": "object",
        "properties": {
          "followers": {
            "type": "number"
          },
          "following": {
            "type": "number"
          },
          "isFollowing": {
            "type": "boolean"
          }
        },
        "required": [
          "followers",
          "following",
          "isFollowing"
        ]
      },
      "FollowUsersPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "displayName": {
                  "type": "string"
                },
                "avatarUrl": {
                  "type": "string"
                },
                "followedAt": {
                  "type": "string"
                }
              },
              "required": [
                "userId",
                "displayName",
                "followedAt"
              ]
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "perPage": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "totalPages": {
                "type": "number"
              },
              "nextCursor": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "page",
              "perPage",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}